refactor: change the package path of MySQL recognizer (#2125)
