+++ rm-datasource/src/test/java/io/seata/rm/datasource/exec/DeleteExecutorTest.java
@@ -32,7 +32,7 @@ import io.seata.rm.datasource.DataSourceProxy;
 import io.seata.rm.datasource.StatementProxy;
 import io.seata.rm.datasource.mock.MockDriver;
 import io.seata.rm.datasource.sql.struct.TableRecords;
-import io.seata.sqlparser.druid.MySQLDeleteRecognizer;
+import io.seata.sqlparser.druid.mysql.MySQLDeleteRecognizer;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
