+++ sqlparser/seata-sqlparser-druid/src/main/resources/META-INF/services/io.seata.sqlparser.druid.SQLOperateRecognizerHolder
@@ -1,2 +1,2 @@
-io.seata.sqlparser.druid.MySqlOperateRecognizerHolder
+io.seata.sqlparser.druid.mysql.MySQLOperateRecognizerHolder
 io.seata.sqlparser.druid.oracle.OracleOperateRecognizerHolder
