+++ sqlparser/seata-sqlparser-druid/src/test/java/io/seata/sqlparser/druid/MySQLSelectForUpdateRecognizerTest.java
@@ -26,6 +26,7 @@ import com.alibaba.druid.sql.ast.statement.SQLSelectStatement;
 import io.seata.sqlparser.ParametersHolder;
 import io.seata.sqlparser.SQLParsingException;
 import io.seata.sqlparser.SQLType;
+import io.seata.sqlparser.druid.mysql.MySQLSelectForUpdateRecognizer;
 import io.seata.sqlparser.util.JdbcConstants;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
